[chore/#32] Int-> Long 으로 변경

+++ app/src/main/java/org/sopt/lotteshopping/data/mapper/HomeTopBannersMapper.kt
@@ -4,7 +4,7 @@ import org.sopt.lotteshopping.data.model.banners.HomeTopBannerModel
 import org.sopt.lotteshopping.data.remote.dto.banners.GetHomeTopBannersResponse
 
 fun GetHomeTopBannersResponse.toModel() = HomeTopBannerModel(
-    id = id.toLong(),
-    seq = seq.toLong(),
+    id = id,
+    seq = seq,
     imageUrl = imageUrl ?: ""
 )
